Rename applycrds pkg to crdutil
